+++ ydb/apps/ydb/CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added ability of `ydb workload tpch` and `ydb workload tpcds` commands to use fraction `--scale` option.
+* Added `ydb workload tpcc check` subcommand, which checks TPC-C data consistency.
 
 ## 2.23.0 ##